+++ src/test/java/com/iexec/sms/secret/web2/Web2SecretServiceTests.java
@@ -22,9 +22,12 @@ import com.iexec.sms.encryption.EncryptionService;
 import com.iexec.sms.secret.MeasuredSecretService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
+import org.springframework.boot.test.system.CapturedOutput;
+import org.springframework.boot.test.system.OutputCaptureExtension;
 
 import java.util.Optional;
 
@@ -32,6 +35,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
 
+@ExtendWith(OutputCaptureExtension.class)
 class Web2SecretServiceTests {
     private static final String OWNER_ADDRESS = "ownerAddress";
     private static final String SECRET_ADDRESS = "secretAddress";
@@ -108,6 +112,49 @@ class Web2SecretServiceTests {
         verify(web2SecretRepository, times(1))
                 .findById(any(Web2SecretHeader.class));
         verifyNoInteractions(encryptionService);
+    }
+    // endregion
+
+
+    // region isSecretPresent
+    @Test
+    void shouldGetSecretExistenceFromDBAndPutInCache(CapturedOutput output) {
+        final Web2Secret encryptedSecret = new Web2Secret(OWNER_ADDRESS, SECRET_ADDRESS, ENCRYPTED_SECRET_VALUE);
+
+        when(web2SecretRepository.findById(any(Web2SecretHeader.class)))
+                .thenReturn(Optional.of(encryptedSecret));
+
+        final boolean resultFirstCall = web2SecretService.isSecretPresent(OWNER_ADDRESS, SECRET_ADDRESS);
+
+        assertAll(
+                () -> assertTrue(resultFirstCall),
+                () -> verify(web2SecretRepository, times(1)).findById(any()),
+                () -> assertThat(output.getOut()).contains("Search secret existence in cache"),
+                () -> assertThat(output.getOut()).contains("Secret existence was not found in cache"),
+                () -> assertThat(output.getOut()).doesNotContain("Secret existence was found in cache"),
+                () -> assertThat(output.getOut()).contains("Put secret existence in cache")
+        );
+    }
+
+    @Test
+    void shouldGetSecretExistenceFromCache(CapturedOutput output) {
+        final Web2Secret encryptedSecret = new Web2Secret(OWNER_ADDRESS, SECRET_ADDRESS, ENCRYPTED_SECRET_VALUE);
+        when(web2SecretRepository.findById(any(Web2SecretHeader.class)))
+                .thenReturn(Optional.of(encryptedSecret));
+
+        web2SecretService.isSecretPresent(OWNER_ADDRESS, SECRET_ADDRESS);
+        final int logLengthForFirstCall = output.getOut().length();
+        final boolean resultSecondCall = web2SecretService.isSecretPresent(OWNER_ADDRESS, SECRET_ADDRESS);
+        final String secondCallLogs = output.getOut().substring(logLengthForFirstCall - 1);
+        assertAll(
+                () -> assertTrue(resultSecondCall),
+                //put 1 bellow means no new invocation since 1st call
+                () -> verify(web2SecretRepository, times(1)).findById(any()),
+                () -> assertThat(secondCallLogs).doesNotContain("Secret existence was not found in cache"),
+                () -> assertThat(secondCallLogs).doesNotContain("Put secret existence in cache"),
+                () -> assertThat(secondCallLogs).contains("Search secret existence in cache"),
+                () -> assertThat(secondCallLogs).contains("Secret existence was found in cache")
+        );
     }
     // endregion